info-overview: reduce size of logo when using icon from /etc/os-release
When using the icon from /etc/os-release, display it as 128x128 rather
than 256x256. In distributions that don't ship a 128x128, such as
Fedora, this results in the 256x256 icon being scaled down to a
reasonable size. 256x256 is so large here as to be clearly undesirable.
128x128 is also the size that Ubuntu uses in its downstream patch. Might
as well reduce the need for patching as far as possible, even though
Fedora doesn't plan to use this codepath.